accel/amdxdna: split the common DRM declarations into amdxdna_drv.h
Move the transport-independent amdxdna declarations (the XDNA_* log macros, the
drm_driver, struct amdxdna_dev / amdxdna_dev_ops / amdxdna_dev_info /
amdxdna_client / amdxdna_dpt_chan, and the iommu/sysfs/dma helpers) out of
amdxdna_pci_drv.h into a new common header amdxdna_drv.h. amdxdna_pci_drv.h now
includes amdxdna_drv.h and keeps only the PCI-only bits: struct
amdxdna_device_id and the PCI dev_*_info externs.

No functional change: files that include amdxdna_pci_drv.h still see everything
via amdxdna_drv.h. This lets the platform driver (aie4_plat.c) and other non-PCI
code include amdxdna_drv.h without pulling in the PCI device-id table.

Ported from 2d076142 (aie4-common), reconstructed on this branch's header
content so it keeps the configurable fw-log buf_size field and the current
dev_*_info names (no dev_info rename).
