+++ drivers/accel/amdxdna/aie.c
@@ -18,7 +18,7 @@
 #include "amdxdna_gem.h"
 #include "amdxdna_mailbox_helper.h"
 #include "amdxdna_mailbox.h"
-#include "amdxdna_pci_drv.h"
+#include "amdxdna_drv.h"
 #include "amdxdna_pm.h"
 
 void aie_dump_mgmt_chann_debug(struct aie_device *aie)
