+++ drivers/accel/amdxdna/aie4.c
@@ -22,7 +22,7 @@
 #include "amdxdna_error.h"
 #include "amdxdna_gem.h"
 #include "amdxdna_mailbox_helper.h"
-#include "amdxdna_pci_drv.h"
+#include "amdxdna_drv.h"
 #include "amdxdna_pm.h"
 #include "amdxdna_sensors.h"
 #include "amdxdna_tile_read_write.h"
