+++ drivers/accel/amdxdna/aie4_pci.c
@@ -27,7 +27,7 @@
 #include "amdxdna_gem.h"
 #include "amdxdna_mailbox.h"
 #include "amdxdna_mailbox_helper.h"
-#include "amdxdna_pci_drv.h"
+#include "amdxdna_drv.h"
 #include "amdxdna_pm.h"
 #include "amdxdna_sensors.h"
 #include "trace/events/amdxdna.h"
