+++ drivers/accel/amdxdna/amdxdna_drv.h
@@ -1,0 +1,297 @@
+/* SPDX-License-Identifier: GPL-2.0 */
+/*
+ * Copyright (C) 2022-2026, Advanced Micro Devices, Inc.
+ *
+ * Transport-independent (common) amdxdna DRM/device declarations, shared by the
+ * PCI driver and the platform driver.  The PCI-only (device, revision) match
+ * table and the PCI device_info externs live in amdxdna_pci_drv.h.
+ */
+
+#ifndef _AMDXDNA_DRV_H_
+#define _AMDXDNA_DRV_H_
+
+#include "drm/amdxdna_accel.h"
+#include <drm/drm_device.h>
+#include <drm/drm_file.h>
+#include <drm/drm_mm.h>
+#include <drm/drm_print.h>
+#include <linux/capability.h>
+#include <linux/cred.h>
+#include <linux/idr.h>
+#include <linux/iommu.h>
+#include <linux/iova.h>
+#include <linux/mm_types.h>
+#include <linux/sched.h>
+#include <linux/srcu.h>
+#include <linux/workqueue.h>
+#include <linux/xarray.h>
+
+#define XDNA_INFO(xdna, fmt, args...)	drm_info(&(xdna)->ddev, fmt, ##args)
+#define XDNA_WARN(xdna, fmt, args...)	drm_warn(&(xdna)->ddev, "%s: "fmt, __func__, ##args)
+#define XDNA_ERR(xdna, fmt, args...)	drm_err(&(xdna)->ddev, "%s: "fmt, __func__, ##args)
+#define XDNA_DBG(xdna, fmt, args...)	drm_dbg(&(xdna)->ddev, fmt, ##args)
+#define XDNA_INFO_ONCE(xdna, fmt, args...) drm_info_once(&(xdna)->ddev, fmt, ##args)
+#define XDNA_ERR_RATELIMITED(xdna, fmt, args...) \
+	drm_err_ratelimited(&(xdna)->ddev, "%s: "fmt, __func__, ##args)
+
+#define XDNA_MBZ_DBG(xdna, ptr, sz)					\
+	({								\
+		int __i;						\
+		int __ret = 0;						\
+		u8 *__ptr = (u8 *)(ptr);				\
+		for (__i = 0; __i < (sz); __i++) {			\
+			if (__ptr[__i]) {				\
+				XDNA_DBG(xdna, "MBZ check failed");	\
+				__ret = -EINVAL;			\
+				break;					\
+			}						\
+		}							\
+		__ret;							\
+	})
+
+#define to_xdna_dev(drm_dev) \
+	((struct amdxdna_dev *)container_of(drm_dev, struct amdxdna_dev, ddev))
+
+extern const struct drm_driver amdxdna_drm_drv;
+
+struct amdxdna_client;
+struct amdxdna_dev;
+struct amdxdna_dev_hdl;
+struct amdxdna_dev_priv;
+struct amdxdna_drm_get_info;
+struct amdxdna_drm_set_state;
+struct amdxdna_gem_obj;
+struct amdxdna_hwctx;
+struct amdxdna_rev_vbnv;
+struct amdxdna_sched_job;
+struct amdxdna_msg_buf_hdl;
+struct aie_device;
+struct aie_error_lut_set;
+
+/*
+ * struct amdxdna_dev_ops - Device hardware operation callbacks
+ */
+struct amdxdna_dev_ops {
+	int (*init)(struct amdxdna_dev *xdna);
+	void (*fini)(struct amdxdna_dev *xdna);
+	void (*debugfs_init)(struct amdxdna_dev *xdna);
+	int (*resume)(struct amdxdna_dev *xdna);
+	int (*suspend)(struct amdxdna_dev *xdna);
+	int (*runtime_resume)(struct amdxdna_dev *xdna);
+	int (*runtime_suspend)(struct amdxdna_dev *xdna);
+	void (*reset_prepare)(struct amdxdna_dev *xdna);
+	void (*reset_done)(struct amdxdna_dev *xdna);
+	int (*sriov_configure)(struct amdxdna_dev *xdna, int num_vfs);
+	int (*mmap)(struct amdxdna_client *client, struct vm_area_struct *vma);
+	int (*hwctx_init)(struct amdxdna_hwctx *hwctx);
+	void (*hwctx_fini)(struct amdxdna_hwctx *hwctx);
+	int (*hwctx_config)(struct amdxdna_hwctx *hwctx, u32 type, u64 value, void *buf, u32 size);
+	int (*hwctx_sync_debug_bo)(struct amdxdna_hwctx *hwctx, u32 debug_bo_hdl);
+	int (*hwctx_heap_expand)(struct amdxdna_hwctx *hwctx, struct amdxdna_gem_obj *heap);
+	int (*cmd_submit)(struct amdxdna_hwctx *hwctx, struct amdxdna_sched_job *job, u64 *seq);
+	int (*cmd_wait)(struct amdxdna_hwctx *hwctx, u64 seq, u32 timeout);
+	int (*get_aie_info)(struct amdxdna_client *client, struct amdxdna_drm_get_info *args);
+	int (*set_aie_state)(struct amdxdna_client *client, struct amdxdna_drm_set_state *args,
+			     u32 *settle_ms);
+	int (*get_array)(struct amdxdna_client *client, struct amdxdna_drm_get_array *args);
+	int (*get_dev_revision)(struct amdxdna_dev *xdna, u32 *rev);
+
+	/* Asynchronous error reporting callbacks. */
+	int (*register_async_event)(struct aie_device *aie, dma_addr_t addr, u32 size,
+				    void *handle, int (*cb)(void *, void __iomem *, size_t));
+	/*
+	 * Handle a device/architecture-specific async event type (beyond the
+	 * generic AIE tile error). Returns true when it consumed the event, so
+	 * the shared AIE tile error decode is skipped; false to let the caller
+	 * run the shared tile decode.
+	 */
+	bool (*handle_dev_async_event)(struct aie_device *aie, u32 type, void *vaddr);
+};
+
+struct amdxdna_fw_feature_tbl {
+	u64 features;
+	u32 major;
+	u32 max_minor;
+	u32 min_minor;
+};
+
+/*
+ * struct amdxdna_dev_info - Device hardware information
+ * Record device static information, like reg, mbox, PSP, SMU bar index
+ */
+struct amdxdna_dev_info {
+	int				reg_bar;
+	int				mbox_bar;
+	int				sram_bar;
+	int				psp_bar;
+	int				smu_bar;
+	int				doorbell_bar;
+	int				device_type;
+	int				first_col;
+	u32				dev_mem_buf_shift;
+	u64				dev_mem_base;
+	size_t				dev_mem_size;
+	const char			*default_vbnv;
+	const struct amdxdna_rev_vbnv	*rev_vbnv_tbl;
+	size_t				dev_heap_max_size;
+	const struct amdxdna_dev_priv	*dev_priv;
+	const struct amdxdna_fw_feature_tbl *fw_feature_tbl;
+	const struct amdxdna_fw_feature_tbl *cert_feature_tbl;
+	const struct amdxdna_dev_ops	*ops;
+
+	/* Asynchronous error reporting data (per AIE generation). */
+	const struct aie_error_lut_set	*luts;
+	u32				async_max_status_code;
+};
+
+struct amdxdna_carveout;
+struct amdxdna_dpt;
+struct amdxdna_dpt_desc;
+struct aie_device;
+
+/*
+ * One firmware Debug/Profile/Trace channel, holding everything needed to
+ * work with it: @srcu guards the lifetime of @data, and @desc carries the
+ * channel's constants and firmware hooks. Callers pass a channel pointer
+ * around rather than a handle plus a discriminator, so the two channels
+ * share one set of helpers without any of them switching on which is which.
+ * struct amdxdna_dpt_desc is private to amdxdna_dpt.c.
+ */
+struct amdxdna_dpt_chan {
+	struct srcu_struct		srcu;
+	struct amdxdna_dpt __rcu	*data;
+	const struct amdxdna_dpt_desc	*desc;
+	/*
+	 * Ring size the next publish asks for, starting at @desc's default.
+	 * Serialized by dev_lock.
+	 */
+	u32				buf_size;
+};
+
+struct amdxdna_dev {
+	struct drm_device		ddev;
+	struct amdxdna_dev_hdl		*dev_handle;
+	const struct amdxdna_dev_info	*dev_info;
+	void				*xrs_hdl;
+
+	struct mutex			dev_lock; /* per device lock */
+	struct list_head		client_list;
+	struct mutex			client_lock; /* client_list */
+	/*
+	 * Device-wide cyclic allocator for hwctx IDs. Only the ID space is shared
+	 * so a context ID is unique across the whole device; the id->hwctx map and
+	 * its lifetime srcu live per-client (see struct amdxdna_client). next_hwctxid
+	 * is the cyclic cursor. Both are serialized by dev_lock.
+	 */
+	struct ida			hwctx_ida;
+	u32				next_hwctxid;
+	struct amdxdna_drm_query_firmware_version fw_ver;
+	struct rw_semaphore		notifier_lock; /* for mmu notifier*/
+	struct workqueue_struct		*notifier_wq;
+
+	struct iommu_group		*group;
+	struct iommu_domain		*domain;
+	struct iova_domain		iovad;
+	/* Accurate board name queried from firmware, or default_vbnv as fallback */
+	const char			*vbnv;
+
+	struct amdxdna_carveout		*carveout;
+
+	/* Firmware Debug/Profile/Trace (DPT) framework. Each channel owns the
+	 * SRCU domain guarding its own handle; on disable we synchronize_srcu
+	 * so kfree of the handle is provably ordered after any watcher's
+	 * srcu_read_unlock.
+	 */
+	struct amdxdna_dpt_chan		fw_log;
+	struct amdxdna_dpt_chan		fw_trace;
+
+	/* Allow auto core dump for hardware contexts, if true. */
+	bool				auto_coredump;
+
+	/* Back-pointer to the per-generation aie_device, set at DPT init so
+	 * common code (e.g. debugfs) can reach msg_ops without knowing the
+	 * generation-specific struct amdxdna_dev_hdl layout.
+	 */
+	struct aie_device		*dpt_aie;
+};
+
+struct amdxdna_io_stats {
+	spinlock_t			lock; /* protect io stats */
+	int				job_depth;
+	u64				start_time;
+	u64				busy_time;
+};
+
+/*
+ * struct amdxdna_client - amdxdna client
+ * A per fd data structure for managing context and other user process stuffs.
+ */
+struct amdxdna_client {
+	struct list_head		node;
+	pid_t				pid;
+	char				name[TASK_COMM_LEN];
+	/*
+	 * Guards hwctx lifetime against this client's own blocking submit/wait.
+	 * A lookup only ever returns this client's own hwctx (per-client xa), so a
+	 * single per-client srcu is sufficient and a parked waiter can only stall
+	 * its own client's teardown.
+	 */
+	struct srcu_struct		hwctx_srcu;
+	/* id->hwctx map for this client; IDs come from xdna->hwctx_ida. */
+	struct xarray			hwctx_xa;
+	struct amdxdna_dev		*xdna;
+	struct drm_file			*filp;
+
+	struct mutex			mm_lock; /* protect memory related */
+	struct list_head		bo_invalid_list; /* protected by xdna->notifier_lock */
+	struct xarray			dev_heap_xa;
+	struct drm_mm			dev_heap_mm;
+	u32				dev_heap_nid;
+	size_t				total_heap_size;
+
+	struct iommu_sva		*sva;
+	int				pasid;
+	struct mm_struct		*mm;
+
+	size_t				heap_usage;
+	size_t				total_bo_usage;
+	size_t				total_int_bo_usage;
+
+	struct amdxdna_io_stats		io_stats;
+};
+
+#define amdxdna_for_each_hwctx(client, hwctx_id, entry)		\
+	xa_for_each(&(client)->hwctx_xa, hwctx_id, entry)
+
+#define amdxdna_for_each_client(xdna, client)			\
+	list_for_each_entry(client, &(xdna)->client_list, node)
+
+int amdxdna_sysfs_init(struct amdxdna_dev *xdna);
+void amdxdna_sysfs_fini(struct amdxdna_dev *xdna);
+
+int amdxdna_iommu_init(struct amdxdna_dev *xdna);
+void amdxdna_iommu_fini(struct amdxdna_dev *xdna);
+void *amdxdna_iommu_alloc(struct amdxdna_dev *xdna, size_t size, dma_addr_t *dma_addr);
+void amdxdna_iommu_free(struct amdxdna_dev *xdna, size_t size,
+			void *cpu_addr, dma_addr_t dma_addr);
+int amdxdna_dma_map_bo(struct amdxdna_dev *xdna, struct amdxdna_gem_obj *abo);
+void amdxdna_dma_unmap_bo(struct amdxdna_dev *xdna, struct amdxdna_gem_obj *abo);
+
+static inline bool amdxdna_iova_on(struct amdxdna_dev *xdna)
+{
+	return !!xdna->domain;
+}
+
+static inline bool amdxdna_pasid_on(struct amdxdna_client *client)
+{
+	return client->pasid != IOMMU_PASID_INVALID;
+}
+
+/* True if the current task may examine @client's contexts. */
+static inline bool amdxdna_client_visible(struct amdxdna_client *client)
+{
+	return capable(CAP_SYS_ADMIN) ||
+	       uid_eq(current_euid(), client->filp->filp->f_cred->euid);
+}
+
+#endif /* _AMDXDNA_DRV_H_ */
