+++ drivers/accel/amdxdna/amdxdna_gem.c
@@ -23,7 +23,7 @@
 #endif
 #include "amdxdna_ctx.h"
 #include "amdxdna_gem.h"
-#include "amdxdna_pci_drv.h"
+#include "amdxdna_drv.h"
 #include "amdxdna_pm.h"
 #include "amdxdna_ubuf.h"
