+++ drivers/accel/amdxdna/amdxdna_iommu.c
@@ -16,7 +16,7 @@
 #endif
 
 #include "amdxdna_gem.h"
-#include "amdxdna_pci_drv.h"
+#include "amdxdna_drv.h"
 
 static bool force_iova;
 module_param(force_iova, bool, 0600);
