+++ drivers/accel/amdxdna/amdxdna_mailbox_helper.c
@@ -14,7 +14,7 @@
 #include "amdxdna_gem.h"
 #include "amdxdna_mailbox.h"
 #include "amdxdna_mailbox_helper.h"
-#include "amdxdna_pci_drv.h"
+#include "amdxdna_drv.h"
 
 int xdna_msg_cb(void *handle, void __iomem *data, size_t size)
 {
