+++ drivers/accel/amdxdna/amdxdna_ubuf.h
@@ -6,7 +6,7 @@
 #define _AMDXDNA_UBUF_H_
 
 #include "amdxdna_gem.h"
-#include "amdxdna_pci_drv.h"
+#include "amdxdna_drv.h"
 
 struct amdxdna_gem_obj *amdxdna_alloc_ubuf_bo(struct amdxdna_client *client,
 					      u32 num_entries, void __user *va_entries);
